Add utility for detecting lingering tasks
This change adds a mechanism for recording the goroutine stack where
Context.Call() and Context.Go() are called in order to detect tasks which did
not respond to the stop signal.

A new "Invoker" type can be attached to a context to observe task creation and
to decorate the task callback function. Invokers compose within a context
hierarchy.

Lingering tasks are tracked by an implementation of Invoker that samples the
goroutine stack at task creation. This data is stored in a sync.Map, which is
indexed by an atomic counter. This is expected to minimize impact on concurrent
operation.

Testable examples for the new APIs are provided.